requirements.txt
asgiref==3.2.3
autopep8==1.5
certifi==2019.11.28
chardet==3.0.4
Django==3.0.3
django-filter==2.2.0
djangorestframework==3.11.0
idna==2.8
Markdown==3.2.1
pycodestyle==2.5.0
pytz==2019.3
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.8
